Separate into assembler and emulator.
