+++ CMakeLists.txt
@@ -1,32 +1,25 @@
 cmake_minimum_required(VERSION 3.28)
-project(glx C)
+project(GameLex C)
 
 set(CMAKE_C_STANDARD 11)
 
+set(GLX_COMMON_SOURCES ./src/text.c ./src/glex.c)
+set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY  ${CMAKE_CURRENT_SOURCE_DIR}/lib)
+add_library(GLX_COMMON_LIB ${GLX_COMMON_SOURCES})
 
-# lib-raylib
-include(FetchContent)
-set(FETCHCONTENT_QUIET FALSE)
-set(BUILD_EXAMPLES OFF CACHE BOOL "" FORCE) # don't build the supplied examples
-set(BUILD_GAMES    OFF CACHE BOOL "" FORCE) # don't build the supplied example games
-FetchContent_Declare(
-        raylib
-        GIT_REPOSITORY "https://github.com/raysan5/raylib.git"
-        GIT_TAG "master"
-        GIT_PROGRESS TRUE
-)
+# toolchain names
+# set(GLX_COMPILER glc)       # Game-Lex-Compiler
+set(GLX_ASSEMBLER glasm)      # Game-Lex-Assembler
+set(GLX_EMULATOR gle)         # Game-Lex-ASM-Emulator
 
-FetchContent_MakeAvailable(raylib)
+set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wall -Wextra -Wswitch-enum -pedantic -std=c11 -ggdb")
 
-# Define PROJECT_SOURCES as a list of all source files
-file(GLOB_RECURSE PROJECT_SOURCES CONFIGURE_DEPENDS "${CMAKE_CURRENT_LIST_DIR}/src/*.c")
-file(GLOB_RECURSE PROJECT_HEADERS CONFIGURE_DEPENDS "${CMAKE_CURRENT_LIST_DIR}/src/*.h")
+set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/bin)
+add_executable(GLX_ASSEMBLER)
+add_executable(GLX_EMULATOR)
 
-add_compile_options(-Wall -Wextra -Wswitch-enum -pedantic -std=c11)
-add_executable(${PROJECT_NAME}
-        src/glx.c
-        src/text.h
-        src/text.c
-)
-target_sources(${PROJECT_NAME} PRIVATE ${PROJECT_SOURCES})
-target_link_libraries(${PROJECT_NAME} PRIVATE raylib)
+target_sources(GLX_ASSEMBLER PRIVATE ./src/glc.c ./src/assembler.c)
+target_sources(GLX_EMULATOR PRIVATE ./src/gle.c ./src/emulator.c)
+
+target_link_libraries(GLX_ASSEMBLER PUBLIC GLX_COMMON_LIB)
+target_link_libraries(GLX_EMULATOR PUBLIC GLX_COMMON_LIB)
