+++ src/glex.c
@@ -1,0 +1,16 @@
+#include <assert.h>
+#include "glex.h"
+
+const char *err_as_cstr(Err err)
+{
+    switch (err) {
+        case ERR_OK: return "ERR_OK";
+        case ERR_STACK_OVERFLOW: return "ERR_STACK_OVERFLOW";
+        case ERR_STACK_UNDERFLOW: return "ERR_STACK_UNDER";
+        case ERR_DIV_BY_ZERO: return "ERR_DIV_BY_ZERO";
+        case ERR_ILLEGAL_INST: return "ERR_ILLEGAL_INST";
+        case ERR_ILLEGAL_INST_ACCESS: return "ERR_ILLEGAL_INST_ACCESS";
+        case ERR_ILLEGAL_OPERAND: return "ERR_ILLEGAL_OPERAND";
+        default:assert(0 && "err_as_cstr: unreachable");
+    }
+}
